only support pythons that are not EOL (https://endoflife.date/python)
Even debian oldstable has python 3.9. For internet-facing libraries it
is not secure for contributor to install unsupported python versions in order to test them.
Reducing the number of python versions will make maintenance and testing easier.
is not secure to contributor to install unsupported python versions in order to test them.